typescript 対応
