+++ package.json
@@ -13,5 +13,10 @@
     "react-dom": "17.0.2",
     "remark": "^14.0.2",
     "remark-html": "^15.0.1"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.5",
+    "@types/react": "^17.0.38",
+    "typescript": "^4.5.4"
   }
 }
