+++ posts/jin.md
@@ -1,0 +1,8 @@
+---
+title: 'jin です'
+date: '2020-01-01'
+---
+
+# あいさつ
+こんにちは、じんです。
+よろしくね
